Added return type to jwt_generator.build_jwt()
Types should be defined wherever possible. This helps users of the SDK know what to expect. User's IDE's/editors will often show what types are returned or expected. The more types there are, the more mypy can validate them in this project and thus validate more of the code in this project. This will also help users of mypy in their own projects that use this SDK to know what types are required.